Remove processTime
We used to make a `setTimeout` to simulate processing.
We discovered that combining setTimeout in the iframe + postMessage +
HTML rendering on the app was causing high latency starting from the
2nd postMessage when +100 docs are generated and transmitted.
Typically, with 100 docs, we measured few ms on the 1st search, and
 +400ms on the 2nd and the next ones.
This behaviour is only seen on Chrome, not on Firefox nor Safari.
This also does not happen when making the `setTimeout` directly in the
web worker.
We disable the setTimeout for now